requirements.txt
certifi==2019.6.16
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Cors==3.0.4
flask-peewee==3.0.0
gunicorn==19.9.0
idna==2.8
iso3166==1.0
iso8601==0.1.12
itsdangerous==0.24
Jinja2==2.10
lxml==4.3.4
MarkupSafe==1.0
peewee==3.3.1
pyactiveresource==2.2.2
pycrypto==2.6.1
PyJWT==2.4.0
PyMySQL==0.8.0
PyYAML==5.2
requests==2.22.0
ShopifyAPI==12.0.0
six==1.11.0
urllib3==1.25.3
Werkzeug==0.14.1
wtf-peewee==3.0.0
WTForms==2.1
